+++ src/lib/components/modal/wasm-verify-submit/index.tsx
@@ -5,8 +5,12 @@ import {
   ModalOverlay,
   useDisclosure,
 } from "@chakra-ui/react";
-import type { MouseEvent, ReactNode } from "react";
+import { useQueryClient } from "@tanstack/react-query";
+import { useCallback } from "react";
+import type { ReactNode } from "react";
 
+import { CELATONE_QUERY_KEYS, useCelatoneApp } from "lib/app-provider";
+import type { WasmVerifyRequest } from "lib/services/types";
 import { useSubmitWasmVerify } from "lib/services/verification/wasm";
 import type { BechAddr32, WasmVerifyStatus } from "lib/types";
 
@@ -25,34 +29,55 @@ interface WasmVerifySubmitModalProps {
 
 interface WasmVerifySubmitModalBodyProps
   extends Omit<WasmVerifySubmitModalProps, "triggerElement"> {
+  onSubmit: (wasmVerifyRequest: WasmVerifyRequest) => void;
+  isLoading: boolean;
+  isSuccess: boolean;
+  isError: boolean;
   onClose: () => void;
 }
 
 const WasmVerifySubmitModalBody = ({
+  isSuccess,
+  isError,
   onClose,
   ...props
 }: WasmVerifySubmitModalBodyProps) => {
-  const { mutate, isLoading, isSuccess, isError } = useSubmitWasmVerify();
-
   if (isError) return <WasmVerifySubmitFailed onClose={onClose} />;
   if (isSuccess) return <WasmVerifySubmitCompleted />;
-  return (
-    <WasmVerifySubmitForm {...props} onSubmit={mutate} isLoading={isLoading} />
-  );
+  return <WasmVerifySubmitForm {...props} />;
 };
 
 export const WasmVerifySubmitModal = ({
+  codeId,
+  codeHash,
+  wasmVerifyStatus,
+  relatedVerifiedCodes,
+  contractAddress,
   triggerElement,
-  ...props
 }: WasmVerifySubmitModalProps) => {
+  const queryClient = useQueryClient();
+  const { currentChainId } = useCelatoneApp();
+  const { mutate, isLoading, isSuccess, isError } = useSubmitWasmVerify();
+
   const { isOpen, onClose, onOpen } = useDisclosure();
+  const handleClose = useCallback(() => {
+    if (isSuccess)
+      queryClient.invalidateQueries({
+        queryKey: [
+          CELATONE_QUERY_KEYS.WASM_VERIFICATION_INFOS,
+          currentChainId,
+          codeId,
+        ],
+      });
+    onClose();
+  }, [codeId, currentChainId, isSuccess, onClose, queryClient]);
 
   return (
     <>
       <Flex
         as="span"
         display="inline-flex"
-        onClick={(e: MouseEvent<HTMLDivElement>) => {
+        onClick={(e) => {
           e.stopPropagation();
           onOpen();
         }}
@@ -61,7 +86,7 @@ export const WasmVerifySubmitModal = ({
       </Flex>
       <Modal
         isOpen={isOpen}
-        onClose={onClose}
+        onClose={handleClose}
         isCentered
         returnFocusOnClose={false}
       >
@@ -71,7 +96,18 @@ export const WasmVerifySubmitModal = ({
           bg="gray.800"
           maxW="100vw"
         >
-          <WasmVerifySubmitModalBody {...props} onClose={onClose} />
+          <WasmVerifySubmitModalBody
+            codeId={codeId}
+            codeHash={codeHash}
+            wasmVerifyStatus={wasmVerifyStatus}
+            relatedVerifiedCodes={relatedVerifiedCodes}
+            contractAddress={contractAddress}
+            onSubmit={mutate}
+            isLoading={isLoading}
+            isSuccess={isSuccess}
+            isError={isError}
+            onClose={handleClose}
+          />
         </ModalContent>
       </Modal>
     </>